Make number of attempts float

+++ src/Repeater.php
@@ -13,10 +13,10 @@ interface Repeater
 {
     /**
      * @param callable $action Callable that may throw a \Throwable.
-     * @param int $times Number of attempts.
+     * @param float $times Number of attempts.
      * @return callable Wrapped $action that will attempt several $times
      *                  before throwing final \Throwable in case of failure.
      * @throws InvalidArgumentException if $times < 1.
      */
-    public function __invoke(callable $action, int $times): callable;
+    public function __invoke(callable $action, float $times): callable;
 }
